sonic/agentcore/requirements.txt
uvicorn[standard]==0.46.0
fastapi==0.136.1
websockets>=14.0
strands-agents[bidi]==1.38.0
pyaudio
requests==2.33.1
boto3>=1.34.0
botocore>=1.43.3
bedrock-agentcore>=1.0.0
